+++ KnowledgeBase.h
@@ -37,7 +37,7 @@ public:
     KnowledgeBase&& operator=(KnowledgeBase&& rhs) = delete;
 
     void Add(const std::string& src_symbol, const std::string& targ_symbol);
-    void Add(uint32_t targ_index, uint32_t src_index);
+    void Add(uint16_t targ_index, uint16_t src_index);
     void ComputeLinkStrengths();
     float GetPercentOfElementsLessThanThreshold(uint32_t threshold);
     std::unique_ptr<IExcitationVector<float>> Transmit(const IExcitationVector<float>& normalized_excitations) const;
